+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/secrets/DatabaseKey.kt
@@ -4,9 +4,9 @@ import de.connect2x.trixnity.core.model.UserId
 
 private const val ID = "de.connect2x.trixnity.messenger.secrets.database"
 
-suspend fun SecretByteArrays.setDatabaseKey(userId: UserId, key: ByteArray) = set("$ID-$userId", key)
+suspend fun SecretByteArrays.setDatabaseKey(userId: UserId, key: ByteArray) = set(SecretId(userId, ID), key)
 
-suspend fun SecretByteArrays.deleteDatabaseKey(userId: UserId) = set("$ID-$userId", null)
+suspend fun SecretByteArrays.deleteDatabaseKey(userId: UserId) = set(SecretId(userId, ID), null)
 
 suspend fun SecretByteArrays.getDatabaseKey(userId: UserId): ByteArray? =
-    get("$ID-$userId") ?: get(ID) // TODO legacy path can be removed in future
+    get(SecretId(userId, ID)) ?: get(ID) // TODO legacy path can be removed in future