+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/secrets/SecretByteArrayManipulationException.kt
@@ -1,0 +1,4 @@
+package de.connect2x.trixnity.messenger.secrets
+
+class SecretByteArrayManipulationException(message: String, cause: Throwable? = null) :
+    IllegalStateException(message, cause)