+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/connecting/MatrixClientInitializationFailureViewModel.kt
@@ -2,7 +2,12 @@ package de.connect2x.trixnity.messenger.viewmodel.connecting
 
 import de.connect2x.trixnity.core.model.UserId
 import de.connect2x.trixnity.messenger.MatrixClientInitializationException
+import de.connect2x.trixnity.messenger.MatrixClientInitializationException.DatabaseAccessException
+import de.connect2x.trixnity.messenger.MatrixClientInitializationException.DatabaseCannotBeDecryptedException
+import de.connect2x.trixnity.messenger.MatrixClientInitializationException.DatabaseKeysManipulatedException
 import de.connect2x.trixnity.messenger.MatrixClients
+import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
+import de.connect2x.trixnity.messenger.multi.ProfileManager
 import de.connect2x.trixnity.messenger.util.CloseApp
 import de.connect2x.trixnity.messenger.util.getOrNull
 import de.connect2x.trixnity.messenger.viewmodel.ViewModelContext
@@ -13,10 +18,15 @@ interface MatrixClientInitializationFailureViewModelFactory {
     fun create(
         viewModelContext: ViewModelContext,
         userId: UserId,
-        exception: MatrixClientInitializationException,
+        initializationException: MatrixClientInitializationException,
         onDeletionFinished: () -> Unit,
     ): MatrixClientInitializationFailureViewModel {
-        return MatrixClientInitializationFailureViewModelImpl(viewModelContext, userId, exception, onDeletionFinished)
+        return MatrixClientInitializationFailureViewModelImpl(
+            viewModelContext,
+            userId,
+            initializationException,
+            onDeletionFinished,
+        )
     }
 
     companion object : MatrixClientInitializationFailureViewModelFactory
@@ -25,31 +35,80 @@ interface MatrixClientInitializationFailureViewModelFactory {
 interface MatrixClientInitializationFailureViewModel {
     val userId: UserId
     val deleteEnabled: Boolean
+    val initializationException: MatrixClientInitializationException?
 
     fun closeApplication()
 
     fun delete()
+
+    /** Confirms that the deletion of the account or profile was successful. */
+    fun confirmDeletion()
 }
 
 open class MatrixClientInitializationFailureViewModelImpl(
     viewModelContext: ViewModelContext,
     override val userId: UserId,
-    exception: MatrixClientInitializationException,
+    override val initializationException: MatrixClientInitializationException,
     private val onDeletionFinished: () -> Unit,
 ) : ViewModelContext by viewModelContext, MatrixClientInitializationFailureViewModel {
 
-    override val deleteEnabled = exception !is MatrixClientInitializationException.DatabaseLockedException
+    override val deleteEnabled = initializationException is DatabaseAccessException
+
+    val matrixClients = get<MatrixClients>()
+    val messengerConfiguration = get<MatrixMessengerConfiguration>()
+    val profileManager = get<ProfileManager>()
+    var deleteProfile = false
+
+    init {
+        when (initializationException) {
+            is DatabaseKeysManipulatedException -> {
+                log.error { "The keys for the database have been tampered with. Will delete the current profile." }
+                // we cannot delete the profile immediately as it instantly removes this view, so we defer it until a
+                // user interaction takes place [confirmDeletion].
+                deleteProfile = true
+            }
+
+            is DatabaseCannotBeDecryptedException -> {
+                log.error {
+                    "The database cannot be decrypted. Will log out automatically and delete all local data for account."
+                }
+                deleteAccountAndWait()
+                // UI should call confirmDeletion() afterwards
+            }
+            else -> {
+                log.debug { "Found init exception: $initializationException" }
+            }
+        }
+    }
 
     override fun closeApplication() {
         getOrNull<CloseApp>()?.invoke()
     }
 
-    val matrixClients = get<MatrixClients>()
-
     override fun delete() {
         coroutineScope.launch {
             matrixClients.remove(userId)
             onDeletionFinished()
+        }
+    }
+
+    override fun confirmDeletion() {
+        if (deleteProfile) {
+            deleteProfile()
+        }
+        onDeletionFinished()
+    }
+
+    private fun deleteAccountAndWait() {
+        coroutineScope.launch { matrixClients.remove(userId) }
+    }
+
+    private fun deleteProfile() {
+        coroutineScope.launch {
+            profileManager.activeProfile.value?.let { profile ->
+                log.warn { "Deleting profile: $profile" }
+                profileManager.deleteProfile(profile)
+            } ?: log.warn { "Tried to delete profile, but no active profile found." }
         }
     }
 }